Move policy checks to calling side of rpc.
This prepares to move some of the functionality in the manager
up to the api. We previously did policy checking on the manager
side because quantum v1 was a manager. Now that quantum replaces
the api, we can do the policy checking on the api side.
