+++ etc/nova/policy.json
@@ -102,15 +102,20 @@
     "volume_extension:volume_admin_actions:force_delete": "rule:admin_api",
 
 
-    "network:get_all_networks": "",
-    "network:get_network": "",
-    "network:delete_network": "",
-    "network:disassociate_network": "",
+    "network:get_all": "",
+    "network:get": "",
+    "network:create": "",
+    "network:delete": "",
+    "network:associate": "",
+    "network:disassociate": "",
     "network:get_vifs_by_instance": "",
     "network:allocate_for_instance": "",
     "network:deallocate_for_instance": "",
     "network:validate_networks": "",
     "network:get_instance_uuids_by_ip_filter": "",
+    "network:get_instance_id_by_floating_address": "",
+    "network:setup_networks_on_host": "",
+    "network:get_backdoor_port": "",
 
     "network:get_floating_ip": "",
     "network:get_floating_ip_pools": "",
@@ -121,6 +126,9 @@
     "network:deallocate_floating_ip": "",
     "network:associate_floating_ip": "",
     "network:disassociate_floating_ip": "",
+    "network:release_floating_ip": "",
+    "network:migrate_instance_start": "",
+    "network:migrate_instance_finish": "",
 
     "network:get_fixed_ip": "",
     "network:get_fixed_ip_by_address": "",
